CASIULTIMO

+++ club/src/main/java/com/club/club/repository/PuntuacionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
-public interface PuntuacionRepository extends JpaRepository<Puntuacion, Long> {
+public interface PuntuacionRepository extends JpaRepository<Puntuacion, Integer> {
 
 }
